// src/interFace/interFace.ts
export interface CartProductType {
  _id: string;
  categoryName: string;
  oldPrice: number;
  price: number;
  productDetails: string;
  productImages: string[];
  productName: string;
  productQuantity: number;
  subcategoryName: string;
  img: string;
  date: string;
  orderData?: string;
  offer: boolean;
  offerPersent: number;
  rettings: number[];
  productStatus: string;
  __v?: number;
  totalCard: number;
  bayerEmail?: string;
  clientIP?: string;
  productId?: string;
  netprice?: number;
  totalPrice?: number;
  retting?: number;
  averageRating: number;
  numRatings?: number;
}